BDIINF_SB-1545: Search Endpoint
Feedback from @ltshb:
- Moved ValidateSearchRequest class to validators_serializer.py
- Removed some redundant error testing from managers.py
- Docstring for new function in utils
